Linking <> Worked on the authentication system

+++ Backend/routes/api.php
@@ -5,6 +5,7 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\UserController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\EmergencyController;
+use App\Http\Middleware\AdminMiddleware;
 
 /*
 |--------------------------------------------------------------------------
@@ -23,13 +24,16 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 
 Route::group(["prefix" => "v0.1"], function(){
     // USER CONTROLLER APIs
+   
     Route::group(["prefix" => "user"], function(){
         Route::post("register", [UserController::class, "register"]);
         Route::post("login", [UserController::class, "login"]);
-        Route::post("logout", [UserController::class, "logout"]);
+        Route::post('logout', [UserController::class, 'logout'])->middleware('auth:api');
+        Route::group(["middleware" =>"auth:sanctum"], function(){
         Route::put("apply", [UserController::class, "apply"]);
         Route::get('{id}/get-user-info', [UserController::class, 'getUserInfo']);
         Route::get("get-user-shifts/{id}", [UserController::class, "getUserShifts"]);
+        Route::get("get-request-status/{id}", [UserController::class, "getRequestStatus"]);
         Route::put('edit-bio', [UserController::class, 'editBio']);
         Route::put('edit-tags', [UserController::class, 'editTags']);
         Route::post('request-cover', [UserController::class, 'requestCover']);
@@ -45,7 +49,9 @@ Route::group(["prefix" => "v0.1"], function(){
         Route::put('add-case-report', [UserController::class, 'addCaseReport']);
         Route::get('get-semester', [UserController::class, 'getSemester']);     
     });
+});
     // EMERGENCY CONTROLLER APIs
+    Route::group(["middleware" =>"auth:sanctum"], function(){
     Route::group(["prefix" => "emergency"], function(){
         Route::post("report-emergency", [EmergencyController::class, "reportEmergency"]);
         Route::get('get-ongoing-emergencies', [EmergencyController::class, 'getOngoingEmergencies']);
@@ -60,8 +66,13 @@ Route::group(["prefix" => "v0.1"], function(){
         Route::post("add-assessment", [EmergencyController::class, "addAssessment"]);
         Route::put('end-emergency', [EmergencyController::class, 'endEmergency']);
     });
+});
     // ADMIN CONTROLLER APIs
+    
+        // Admin routes
+   
     Route::group(["prefix" => "admin"], function(){
+        Route::group(["middleware" => ["auth:sanctum", "App\Http\Middleware\AdminMiddleware"]], function() {
         Route::put("update-semester-dates", [AdminController::class, "updateSemesterDates"]);    
         Route::put("add-member", [AdminController::class, "addMember"]);    
         Route::put("remove-member", [AdminController::class, "removeMember"]);    
@@ -69,7 +80,7 @@ Route::group(["prefix" => "v0.1"], function(){
         Route::put("change-rank", [AdminController::class, "changeRank"]);     
         Route::get("get-user-shifts/{id}", [AdminController::class, "getUserShifts"]);
         Route::post("add-shift", [AdminController::class, "addShift"]);    
-        Route::delete("delete-shift", [AdminController::class, "deleteShift"]);   
+        Route::delete("delete-shift/{shift_id}/{user_id}", [AdminController::class, "deleteShift"]);   
         Route::post("add-announcement", [AdminController::class, "addAnnouncement"]);   
         Route::delete("delete-announcement", [AdminController::class, "deleteAnnouncement"]);
         Route::post("add-faq", [AdminController::class, "addFaq"]);
@@ -85,5 +96,6 @@ Route::group(["prefix" => "v0.1"], function(){
         Route::get('get-ongoing-shifts', [AdminController::class, 'getOngoingShifts']);
         Route::get('get-admins', [AdminController::class, 'getAdmins']);
     });
+});
 });
